(refactor: weather-mid-component param 수정)
